refactor: Constraints.fixedCopy extension 이동
